requirements.txt
tensorflow==2.13.0
opencv-python==4.10.0.84
numpy==1.24.0
pillow==9.4.0
streamlit==1.25.0
matplotlib